Let swayfx-git depend on scenefx-git
The swayfx git package depends on specific version of scenefx.
While this generally works fine. This breaks whenever scenefx-git is
installed. Because the provided library may change at any time. Causing
issues with swayfx-git still depending on that specific version.

To solve this depend on the scenefx-git package instead of the provided
library version. This fixes the issue where scenefx-git may install an updated
library version that swayfx-git depends on in the same update cycle.

The only potential risk with this change is that swayfx-git will break
when scenefx-git updates it's library version while not updating
swayfx-git. This results in a runtime issue where swayfx-git can't find
a specific version of the scenefx library.

ref: https://github.com/WillPower3309/swayfx/issues/384

Signed-off-by: Darrel Griët <dgriet@gmail.com>